Stylize+Update MusicStore.java
This was mainly an alteration of parentheses and double spacing for readability, but I did also alter the use of albumCpy for Song, so I went back and fixed this Class to handle that change. I saw no use cases for each Song having its own access to an entire Album with Song Copies in it - it felt minorly overkill - and it also impeded ease of testing. Hope that's ok!