Altered Song.java 's albumCpy usage
This is the original change that spurred on the changes in MusicStore. albumCpy itself still exists as that is useful for assuring that encapsulation is kept, but Songs themselves do not need access to the entire Album. Also, it helps with SongTester, which I will upload once I have a few other Test classes done too.