Half complete Library Test file!
Only 5 regular Tests and 5 EMPTY tests left. I also need to remember to double-check that the isInLibrary methods are fully tested, including all if branching. I have a stand-up desk and my hip hurts and I haven't eaten dinner though, so this is it for the night. Ow. Hopefully I'll wrap up tomorrow, leaving us with the Access, UML Diagram, Essays, Video, and turning this in.