Fixed addAlbum song dupes, made albumBySong()
addAlbum() now correctly does not add duplicate Songs if those songs had already been added prior to the album, and albumBySong() now outputs the album info of the complete album in the Store along with if it is contained in the Library or not. getSize() was also added, and this is a very simple method to get the size of an album contained in the Library (just so albumBySong() can see if its the right length compared to the "full" album in the Store).