Added support for seeing the album info
If a Song in the Library is searched for, the program now provides the option to search for the album info related to their search (whether this be by title or by artist!). This program will never support searching for this information if its searching in the Store, as this is not only redundant, but the spec does not require this.